update turtlegrapics to use ScalaSpecInputListing
