+++ compendium/modules/w06-lab.tex
@@ -22,6 +22,7 @@
 
 Under den här laborationen ska du skriva en samling av klasser som tillsammans kan användas för att rita i en fönsterruta. För att ge dig en grund att stå på får du tillgång till den färdigskrivna klassen SimpleWindow. SimpleWindow är en fönsterruta som tillåter programmeraren att rita linjer med hjälp utav en  ''penna''. SimpleWindow håller koll på pennans nuvarande position och kan ombes att flytta pennan genom att antingen lyfta den eller genom att rita en rak linje till en ny position.
 
+
 \begin{JavaSpec}{class SimpleWindow}
   /** mouse click event type */
 	public final static int MOUSE\_EVENT = 1;
@@ -104,77 +105,13 @@ Med hjälp utav SimpleWindow kan vi nu skapa en Turtle-klass som fungerar likt d
 
 \Task Skapa en klass Point för att beskriva en viss koordinat (x,y) i ett fönster. Klassen ska vara omuterbar - man ska alltså inte kunna ändra på en koordinat efter att den har skapats. Notera att klassens attribut är av typen Double och inte Int, trots att de i någon mån beskriver en diskret pixelposition. Anledningen till detta är att det kan uppstå avrundningsfel vid upprepade förflyttningar. Detta blir särskilt märkbart då varje förflyttning är liten, som t.ex. när en Turtle används för att rita en cirkel.
 
-\begin{ScalaSpec}{Point}
-/** Represents a single Point in the x,y plane. */
-case class Point(val x: Double, val y: Double) {
-
-** Returns a new Point which has been moved some number of pixels */
-def translate(dx: Double, dy: Double) = ???
-}
-\end{ScalaSpec}
+\ScalaSpecInputListing{Point}{../workspace/w06_turtlegraphics/src/main/scala/turtlegraphics/Point.scala}
 
 \Subtask Implementera klassen Point.
 
-
-
-
 \Task Skapa klassen Turtle:
 
-\begin{ScalaSpec}{Turtle}
-/** A Kojo-like Turtle class that can be used to draw shapes
-    in a SimpleWindow.
-  * @param window     The window the turtle should be placed in.
-  * @param position   A Point representing the turtle's starting
-                      coordinates.
-  * @param angle      The angle between the turtle direction and
-                      the X-axis measured in degrees. 
-                      Positive degrees indicate a counter clockwise rotation.
-  * @param isPenDown  A boolean representing the turtle's pen
-                      position. True if the pen is down.
-  */
-class Turtle(window: SimpleWindow, private var position: Point,
-      private var angle: Double, private var isPenDown: Boolean) {
-
-  /** Gets the Turtle's current pixel position on the x axis */
-  def x: Int = ???
-
-  /** Gets the Turtle's current pixel position on the y axis
-      (measured from the top left) */
-  def y: Int = ???
-
-  /** Moves the turtle to a new position without drawing a line. */
-  def jumpTo(newPosition: Point) : Unit = ???
-
-  /** Moves the turtle forward in its current direction, drawing
-      a line if the pen is down.
-    * @param length The number of pixels to move forward.  */
-  def forward(length: Double): Unit = ???
-
-  /** Turns the turtle to the left.
-    *
-    * @param turnAngle The number of degrees to turn. */
-  def turnLeft(turnAngle: Double): Unit = ???
-
-  /** Turns the turtle to the right.
-    *
-    * @param turnAngle The number of degrees to turn. */
-  def turnRight(turnAngle: Double): Unit = ???
-
-  /** Turns the turtle straight up. */
-  def turnNorth(): Unit = ???
-
-  /** Sets the turtle's pen down, causing it to draw lines when
-      the turtle moves. */
-  def penDown(): Unit = ???
-
-  /** Lifts the turtle's pen, preventing it from drawing lines,
-      even when it moves. */
-  def penUp(): Unit = ???
-}
-
-\end{ScalaSpec}
-
-
+\ScalaSpecInputListing{Turtle}{../workspace/w06_turtlegraphics/src/main/scala/turtlegraphics/Turtle.scala}
 
 \Subtask Vilka attribut finns i klassen, och vilken synlighetsnivå har de? Vilken/vilka konstruktorer finns? 
 
@@ -195,35 +132,7 @@ class Turtle(window: SimpleWindow, private var position: Point,
 
 \Task Skapa klassen Rectangle: 
 
-\begin{ScalaSpec}{Rectangle}
-/** Immutable class representing a rectangle.
-  * @param position a Point representing the upper left corner of the
-                    rectangle (before rotation)
-  * @param width    the width of the rectangle
-  * @param height   the height of the rectangle
-  * @param angle    the angle of the rectangle (rotated around
-                    the upper left corner)
-  */
-class Rectangle(position: Point, width: Double,
-                height : Double, angle : Double) {
-  /** Draws the rectangle using a turtle */
-  def draw(turtle: Turtle): Unit = ???
-
-  /** Returns a new Rectangle that is rotated to the left */
-  def rotateLeft(degrees: Double): Rectangle = ???
-
-  /** Returns a new Rectangle that is rotated to the right */
-  def rotateRight(degrees: Double): Rectangle = ???
-
-  /** Returns a new Rectangle that has been scaled by a size factor */
-  def scale(factor: Double): Rectangle = ???
-
-  /** Returns a new Rectangle that has been moved some number of pixels */
-  def translate(dx: Double, dy: Double): Rectangle = ???
-}
-\end{ScalaSpec}
-
-
+\ScalaSpecInputListing{Rectangle}{../workspace/w06_turtlegraphics/src/main/scala/turtlegraphics/Rectangle.scala}
 
 \Subtask Vilken synlighetsnivå bör konstruktorparametrarna ges? Motivera.
 
@@ -254,42 +163,7 @@ while(true){
 \Task Skapa en klass RectangleSequence. I denna klass skall draw-metoden rita ut ett antal rektanglar där varje rektangel har förflyttat sig, roterats och skalats jämfört med föregående rektangel i sekvensen. Se bilder nedan.
 
 
-\begin{ScalaSpec}{RectangleSequence}
-/** Represents a sequence of Rectangles that have been translated,
-    rotated, and scaled down.
-  *
-  * @param rectangle        the rectangle to use as a base for the shape
-  * @param count            the number of rectangles to draw
-  * @param startAngle       the number of degrees to rotate the image
-  * @param step             the number of pixels to move each rectangle
-                            (in the direction of startAngle)
-  * @param rotationStep     the number of degrees to shift each rectangle
-                            with each roll
-  * @param scaleStep        the scale factor to use in each step
-                          
-  */
-case class RectangleSequence(rectangle: Rectangle,
-                             count: Int,
-                             angle: Double,
-                             step: Double,
-                             rotationStep: Double) {
-
-  /** Draws the image using a given Turtle */
-  def draw(turtle: Turtle): Unit = ???
-
-  /** Returns a new RectangleSequence that has been scaled with a size factor*/
-  def scale(factor: Double): RectangleSequence = ???
-
-  /** Returns a new RectangleSequence that has been rotated to the left */
-  def rotateLeft(degrees: Double): RectangleSequence = ???
-
-  /** Returns a new RectangleSequence that has been rotated to the right */
-  def rotateRight(degrees: Double): RectangleSequence = ???
-
-  /** Returns a new RectangleSequence that has moved some number of pixels*/
-  def translate(dx: Double, dy: Double): RectangleSequence = ???
-}
-\end{ScalaSpec}
+\ScalaSpecInputListing{RectangleSequence}{../workspace/w06_turtlegraphics/src/main/scala/turtlegraphics/RectangleSequence.scala}
 
 \Subtask Implementera RectangleSequence.
