+++ workspace/w06_turtlegraphics/src/main/scala/turtlegraphics/Main.scala
@@ -1,0 +1,5 @@
+package turtlegraphics
+
+object Main {
+  def main(args: Array[String]): Unit = ???
+}
