+++ workspace/w06_turtlegraphics/src/main/scala/turtlegraphics/Point.scala
@@ -1,0 +1,9 @@
+package turtlegraphics
+
+/** Represents a single Point in the x,y plane. */
+case class Point(val x: Double, val y: Double) {
+  /** Returns a new Point which has been moved some number of pixels */
+  def translate(dx: Double, dy: Double) = ???
+}
+
+
